+++ tests/e2e/test_spyre_async_llm.py
@@ -2,13 +2,14 @@ import asyncio
 from contextlib import ExitStack
 
 import pytest
-from llm_cache import DecodeWarmupShapes
 from spyre_util import get_chicken_soup_prompts
 from vllm import PromptType, SamplingParams
 from vllm.engine.arg_utils import AsyncEngineArgs
 from vllm.engine.async_llm_engine import AsyncLLMEngine
 from vllm.sampling_params import RequestOutputKind
 from vllm.v1.engine.async_llm import AsyncLLM
+
+from tests.spyre_util import DecodeWarmupShapes
 
 
 async def generate(
@@ -48,10 +49,14 @@ async def generate(
 @pytest.mark.parametrize(
     "output_kind", [RequestOutputKind.DELTA, RequestOutputKind.FINAL_ONLY])
 @pytest.mark.asyncio
-async def test_abort(model: str, backend: str, cb: int,
-                     warmup_shapes: DecodeWarmupShapes,
-                     output_kind: RequestOutputKind,
-                     monkeypatch: pytest.MonkeyPatch):
+async def test_abort(
+    model: str,
+    backend: str,
+    cb: int,
+    warmup_shapes: DecodeWarmupShapes,
+    output_kind: RequestOutputKind,
+    monkeypatch: pytest.MonkeyPatch,
+):
     """Test handling of cancelled requests"""
     with monkeypatch.context() as m, ExitStack() as after:
         m.setenv("VLLM_SPYRE_DYNAMO_BACKEND", backend)
@@ -62,12 +67,18 @@ async def test_abort(model: str, backend: str, cb: int,
             warmup_new_tokens = [t[1] for t in warmup_shapes]
             warmup_batch_size = [t[2] for t in warmup_shapes]
 
-            m.setenv('VLLM_SPYRE_WARMUP_PROMPT_LENS',
-                     ','.join(str(val) for val in warmup_prompt_length))
-            m.setenv('VLLM_SPYRE_WARMUP_NEW_TOKENS',
-                     ','.join(str(val) for val in warmup_new_tokens))
-            m.setenv('VLLM_SPYRE_WARMUP_BATCH_SIZES',
-                     ','.join(str(val) for val in warmup_batch_size))
+            m.setenv(
+                "VLLM_SPYRE_WARMUP_PROMPT_LENS",
+                ",".join(str(val) for val in warmup_prompt_length),
+            )
+            m.setenv(
+                "VLLM_SPYRE_WARMUP_NEW_TOKENS",
+                ",".join(str(val) for val in warmup_new_tokens),
+            )
+            m.setenv(
+                "VLLM_SPYRE_WARMUP_BATCH_SIZES",
+                ",".join(str(val) for val in warmup_batch_size),
+            )
 
         # Async LLM API is a little different between v0 and V1
         engine = AsyncLLM.from_engine_args(
