+++ tests/e2e/test_spyre_warmup_shapes.py
@@ -4,19 +4,24 @@ Run `python -m pytest tests/e2e/test_spyre_warmup_shapes.py`.
 """
 
 import pytest
-from llm_cache import DecodeWarmupShapes
 from spyre_util import (check_output_against_hf, generate_spyre_vllm_output,
                         get_chicken_soup_prompts)
 from vllm import SamplingParams
+
+from tests.spyre_util import DecodeWarmupShapes
 
 
 @pytest.mark.parametrize(
     "warmup_shapes", [[(64, 20, 4),
                        (128, 20, 2)]])  # (prompt_length/new_tokens/batch_size)
-def test_multiple_warmup_shapes(model: str, warmup_shapes: DecodeWarmupShapes,
-                                backend: str, monkeypatch: pytest.MonkeyPatch,
-                                use_llm_cache) -> None:
-    '''
+def test_multiple_warmup_shapes(
+    model: str,
+    warmup_shapes: DecodeWarmupShapes,
+    backend: str,
+    monkeypatch: pytest.MonkeyPatch,
+    use_llm_cache,
+) -> None:
+    """
     The warmup is based on two shapes, that 'overlap' each
     other. After the warmup, one request with the provided
     prompts is input to vLLM. There should be at least one
@@ -29,7 +34,7 @@ def test_multiple_warmup_shapes(model: str, warmup_shapes: DecodeWarmupShapes,
     The same prompts are also input to HF. The generated
     output including text, token ids, and logprobs, is
     verified to be identical for vLLM and HF.
-    '''
+    """
 
     prompts = get_chicken_soup_prompts(4)
 
@@ -39,7 +44,8 @@ def test_multiple_warmup_shapes(model: str, warmup_shapes: DecodeWarmupShapes,
         max_tokens=max_new_tokens,
         temperature=0,
         logprobs=0,  # return logprobs of generated tokens only
-        ignore_eos=True)
+        ignore_eos=True,
+    )
 
     vllm_results = generate_spyre_vllm_output(
         model=model,
@@ -49,7 +55,8 @@ def test_multiple_warmup_shapes(model: str, warmup_shapes: DecodeWarmupShapes,
         sampling_params=vllm_sampling_params,
         tensor_parallel_size=1,
         backend=backend,
-        monkeypatch=monkeypatch)
+        monkeypatch=monkeypatch,
+    )
 
     check_output_against_hf(model, backend, max_new_tokens, vllm_results,
                             prompts)
@@ -57,14 +64,18 @@ def test_multiple_warmup_shapes(model: str, warmup_shapes: DecodeWarmupShapes,
 
 @pytest.mark.parametrize("prompts", [["Hello"]])
 @pytest.mark.parametrize("warmup_shapes", [[(65, 1, 1)]])
-def test_invalid_prompt_len(model: str, prompts: list[str],
-                            warmup_shapes: DecodeWarmupShapes, backend: str,
-                            monkeypatch: pytest.MonkeyPatch,
-                            use_llm_cache) -> None:
-    '''
+def test_invalid_prompt_len(
+    model: str,
+    prompts: list[str],
+    warmup_shapes: DecodeWarmupShapes,
+    backend: str,
+    monkeypatch: pytest.MonkeyPatch,
+    use_llm_cache,
+) -> None:
+    """
     Expects an error to be raised if the warmup prompt length
     is not divisible by 64.
-    '''
+    """
 
     vllm_sampling_params = SamplingParams(max_tokens=1,
                                           temperature=0,
@@ -72,11 +83,13 @@ def test_invalid_prompt_len(model: str, prompts: list[str],
                                           ignore_eos=True)
 
     with pytest.raises(RuntimeError, match="VLLM_SPYRE_WARMUP_PROMPT_LENS"):
-        generate_spyre_vllm_output(model=model,
-                                   prompts=prompts,
-                                   warmup_shapes=warmup_shapes,
-                                   max_model_len=2048,
-                                   sampling_params=vllm_sampling_params,
-                                   tensor_parallel_size=1,
-                                   backend=backend,
-                                   monkeypatch=monkeypatch)
+        generate_spyre_vllm_output(
+            model=model,
+            prompts=prompts,
+            warmup_shapes=warmup_shapes,
+            max_model_len=2048,
+            sampling_params=vllm_sampling_params,
+            tensor_parallel_size=1,
+            backend=backend,
+            monkeypatch=monkeypatch,
+        )
